requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
PyJWT==2.12.1
requests==2.33.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
